requirements.txt
aiogram==3.3.0
pytz
aiohttp
python-dotenv